pyproject.toml
[project]
name = "agent-template"
version = "0.1.0"
description = "Generic chatbot agent with extensible tool architecture"
authors = [
    {name = "Daniel Scholl", email = "dascholl@microsoft.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
keywords = ["agent", "chatbot", "ai", "llm", "tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # Microsoft Agent Framework (latest beta)
    "agent-framework>=1.0.0b251105",

    # Core dependencies
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "prompt-toolkit>=3.0.0",
    "typer>=0.12.0",

    # LLM Provider SDKs (required by agent-framework)
    "openai>=1.58.1",
    "anthropic>=0.40.0",
    "azure-identity>=1.25.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.10.0",
    "black>=25.1.0",
    "mypy>=1.14.1",
]

[project.scripts]
agent = "agent.cli:app"

[project.urls]
Homepage = "https://github.com/danielscholl/agent-template"
Repository = "https://github.com/danielscholl/agent-template"
Issues = "https://github.com/danielscholl/agent-template/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
markers = [
    "integration: Integration tests",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
strict_equality = true

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["agent"]

[tool.coverage.run]
source = ["src/agent"]
omit = [
    "tests/*",
    "src/agent/__main__.py",      # Wrapper only
    "src/agent/utils/__init__.py", # Imports only
    "src/agent/utils/errors.py",  # Exception definitions
    "src/agent/cli.py",           # CLI display logic (hard to test)
    "src/agent/display/*",        # Display logic hard to test
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
